# ⚡️ Make many 400% faster by using while instead of recurse